requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
python-frontmatter==1.1.0
pyyaml==5.1
requests==2.32.3
toml==0.10.2
tqdm==4.67.0
urllib3==2.2.3
